[refactoring] separate dev and prod config
